requirements.txt
ollama
numpy
psycopg2-binary
pgvector
rank_bm25
sentence-transformers
arize-phoenix 
openinference-instrumentation
